+++ Cargo.toml
@@ -20,6 +20,7 @@ members = [
   # implementations
   "hopr/reference",
   "impls/connector",
+  "impls/session-server-forwarder",
   "impls/ct/immediate",
   "impls/ct/full-network",
   "impls/graph",
@@ -180,6 +181,7 @@ hopr-chain-connector = { path = "impls/connector" }
 
 # referential implementations
 hopr-reference = { path = "hopr/reference", default-features = false }
+hopr-session-server-forwarder = { path = "impls/session-server-forwarder", default-features = false }
 hopr-transport-p2p = { path = "impls/transport/p2p", default-features = false }
 hopr-ct-immediate = { path = "impls/ct/immediate", default-features = false }
 hopr-ct-full-network = { path = "impls/ct/full-network", default-features = false }
