+++ impls/session-server-forwarder/Cargo.toml
@@ -1,0 +1,38 @@
+[package]
+name = "hopr-session-server-forwarder"
+description = "HOPR session server that bridges TCP/UDP sockets from the Session Exit node to a destination"
+version = "0.1.0"
+license = "GPL-3.0-only"
+publish = false
+rust-version.workspace = true
+authors.workspace = true
+edition.workspace = true
+
+[features]
+default = ["runtime-tokio"]
+runtime-tokio = [
+  "dep:tokio",
+  "dep:tokio-retry",
+  "hopr-lib/runtime-tokio",
+]
+telemetry = [
+  "hopr-lib/telemetry",
+  "hopr-api/types-telemetry",
+  "dep:lazy_static",
+]
+
+[dependencies]
+async-trait = { workspace = true }
+hopr-api = { workspace = true }
+hopr-lib = { workspace = true, features = ["session-server", "utils_parallel"] }
+lazy_static = { workspace = true, optional = true }
+serde = { workspace = true }
+serde_with = { workspace = true }
+smart-default = { workspace = true }
+tokio = { workspace = true, optional = true }
+tokio-retry = { workspace = true, optional = true }
+tracing = { workspace = true }
+validator = { workspace = true }
+
+[lib]
+doctest = false
